TGP-1413: changed supplementaryUnit data type

+++ app/uk/gov/hmrc/tradergoodsprofilesrouter/models/request/eis/accreditationrequests/GoodsItem.scala
@@ -23,7 +23,7 @@ case class GoodsItem(
   traderReference: String,
   goodsDescription: String,
   countryOfOrigin: Option[String],
-  supplementaryUnit: Option[Int],
+  supplementaryUnit: Option[BigDecimal],
   category: Option[Int],
   measurementUnitDescription: Option[String],
   commodityCode: String
